Rename reception pages
As we're about to allow tube import from SS, we should
rename the reception pages. This change is cosmetic only at
this stage.
